feat: add xor operator
