+++ src/main/scala/satify/dsl/DSL.scala
@@ -18,10 +18,13 @@ object DSL:
 
     def not: Expression = Not(exp1)
 
+    def xor(exp2: Expression): Expression = Or(And(exp1, Not(exp2)), And(Not(exp1), exp2))
+
     @targetName("andSymbol")
     def /\(exp2: Expression): Expression = and(exp2)
     @targetName("orSymbol")
     def \/(exp2: Expression): Expression = or(exp2)
-
     @targetName("notSymbol")
     def unary_! : Expression = not
+    @targetName("xorSymbol")
+    def ⊕(exp2: Expression): Expression = xor(exp2)
