+++ src/test/scala/satify/dsl/MathOperatorsTest.scala
@@ -18,24 +18,8 @@ class MathOperatorsTest extends AnyFlatSpec:
   """ !"A" """ should """ be equal to not("A") """ in {
     !"A" shouldBe not("A")
   }
-/*
-  """ operators """ should "be able to be combined with parenthesis" in {
-    val exp1 = not("A") and "B" or "C"
-    exp1 shouldBe Or(And(Not(Symbol("A")), Symbol("B")), Symbol("C"))
-    val exp2 = "A" and (not("B") or "C")
-    exp2 shouldBe And(Symbol("A"), Or(Not(Symbol("B")), Symbol("C")))
-    val exp3 = ("A" and "B") or not("C")
-    exp3 shouldBe Or(And(Symbol("A"), Symbol("B")), Not(Symbol("C")))
+
+  """ "A" ⊕ "B" """ should """ be equal to "A" xor "B" """ in {
+    "A" ⊕ "B" shouldBe ("A" xor "B")
   }
 
-  """ operators """ should "be left associative" in {
-    val exp1 = "A" and "B" or "C" and not("D")
-    val exp2 = (("A" and "B") or "C") and not("D")
-    val expected = And(
-      Or(And(Symbol("A"), Symbol("B")), Symbol("C")),
-      Not(Symbol("D"))
-    )
-    exp1 shouldBe expected
-    exp2 shouldBe expected
-  }
- */
